Define exceptions in separate module
