+++ py-polars/src/dataframe/mod.rs
@@ -7,7 +7,7 @@ use polars::prelude::*;
 use pyo3::prelude::*;
 use pyo3::types::PyDict;
 
-use crate::PyPolarsErr;
+use crate::error::PyPolarsErr;
 
 #[pyclass]
 #[repr(transparent)]
