+++ py-polars/src/functions/lazy.rs
@@ -7,10 +7,11 @@ use pyo3::types::{PyBool, PyBytes, PyFloat, PyInt, PyString};
 
 use crate::conversion::any_value::py_object_to_any_value;
 use crate::conversion::{get_lf, Wrap};
+use crate::error::PyPolarsErr;
 use crate::expr::ToExprs;
 use crate::map::lazy::binary_lambda;
 use crate::prelude::{vec_extract_wrapped, ObjectValue};
-use crate::{map, PyDataFrame, PyExpr, PyLazyFrame, PyPolarsErr, PySeries};
+use crate::{map, PyDataFrame, PyExpr, PyLazyFrame, PySeries};
 
 macro_rules! set_unwrapped_or_0 {
     ($($var:ident),+ $(,)?) => {
